jl137: casting CERTVerfiyLog.arg, which is void* not working on Linux x86_64

+++ xmlsecurity/source/xmlsec/nss/secerror.cxx
@@ -73,7 +73,7 @@ getCertError(PRErrorCode errNum)
 void
 printChainFailure(CERTVerifyLog *log)
 {
-    unsigned int       errorFlags  = 0;
+    unsigned long errorFlags  = 0;
     unsigned int       depth  = (unsigned int)-1;
     const char * specificError = NULL;
     const char * issuer = NULL;
@@ -98,7 +98,7 @@ printChainFailure(CERTVerifyLog *log)
             switch (node->error)
             {
             case SEC_ERROR_INADEQUATE_KEY_USAGE:
-                errorFlags = (unsigned int)node->arg;
+                errorFlags = (unsigned long)node->arg;
                 switch (errorFlags)
                 {
                 case KU_DIGITAL_SIGNATURE:
@@ -115,7 +115,7 @@ printChainFailure(CERTVerifyLog *log)
                     break;
                 }
             case SEC_ERROR_INADEQUATE_CERT_TYPE:
-                errorFlags = (unsigned int)node->arg;
+                errorFlags = (unsigned long)node->arg;
                 switch (errorFlags)
                 {
                 case NS_CERT_TYPE_SSL_CLIENT:
